% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-regions.R
\name{get_region_codes}
\alias{get_region_codes}
\title{Retrive region names and codes}
\usage{
get_region_codes()
}
\value{
a data frame of region codes and names. The codes can be used
as parameters in other functions that take region codes.
}
\description{
Retrive region names and codes
}
