% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sounding-data.R
\name{get_sounding_data}
\alias{get_sounding_data}
\title{Get sounding data for a station}
\usage{
get_sounding_data(station_number, date, include_frost = FALSE,
  region = c("naconf", "samer", "pac", "nz", "ant", "np", "europe", "africa",
  "seasia", "mideast"), from_hr = c("00", "12", "all"), to_hr = c("00",
  "12", "all"))
}
\arguments{
\item{station_number}{the station number}

\item{date}{an ISO character string (e.g. \code{YYYY-mm-dd}) or a valid \code{Date} object}

\item{include_frost}{if \code{TRUE}, include frost point calculations. Default: \code{FALSE}}

\item{region}{one of "\code{naconf}", "\code{samer}", "\code{pac}", "\code{nz}", "\code{ant}", "\code{np}",
"\code{europe}", "\code{africa}", "\code{seasia}", "\code{mideast}" (which matches the
values of the drop-down menu on the site). Use \code{\link[=get_region_codes]{get_region_codes()}}
to get a mapping of region names to valid region codes.}

\item{from_hr, to_hr}{one of \code{00} (or \code{0}), \code{12} or \code{all}; if \code{all} then both
values will be set to \code{all}}
}
\value{
data frame of soundings data with a "\code{meta}" attribute that contains
metadata about the station and result. Also, the date and from/to hours
are added columns to make it easier to save and reuse the data.
}
\description{
Get sounding data for a station
}
