% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-map.R
\name{get_sounding_map}
\alias{get_sounding_map}
\title{Retrieve pre-generated map for sounding data}
\usage{
get_sounding_map(station_number, date, map_type = c("skewt", "stuve",
  "stuve10", "stuve700", "hodo"), map_format = c("pdf", "gif"),
  include_frost = FALSE, region = c("naconf", "samer", "pac", "nz", "ant",
  "np", "europe", "africa", "seasia", "mideast"), from_hr = c("00", "12",
  "all"), to_hr = c("00", "12", "all"))
}
\arguments{
\item{station_number}{the station number}

\item{date}{an ISO character string (e.g. \code{YYYY-mm-dd}) or a valid \code{Date} object}

\item{map_type}{one of "\code{skewt}" for Skew-T, "\code{stuve}" for Stuve,
"\code{stuve10}" for Stuve to 10 mb, "\code{stuve700}" for Stuve to 700 mb,
"\code{hodo}" for Hodograph. Defaults to "\code{skewt}"}

\item{map_format}{one of "\code{gif}" or "\code{pdf}". Default is "\code{gif}"}

\item{include_frost}{if \code{TRUE}, include frost point calculations. Default: \code{FALSE}}

\item{region}{one of "\code{naconf}", "\code{samer}", "\code{pac}", "\code{nz}", "\code{ant}", "\code{np}",
"\code{europe}", "\code{africa}", "\code{seasia}", "\code{mideast}" (which matches the
values of the drop-down menu on the site). Use \code{\link[=get_region_codes]{get_region_codes()}}
to get a mapping of region names to valid region codes.}

\item{from_hr, to_hr}{one of \code{00} (or \code{0}), \code{12} or \code{all}; if \code{all} then both
values will be set to \code{all}}
}
\value{
a \code{magick} object containing the \code{gif} or \code{pdf} requested.
}
\description{
Retrieve pre-generated map for sounding data
}
\note{
the PDF may be blank or the result may be \code{NULL} (if a GIF was requested)
if there is no data for a given day.
}
