% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind-soundings.R
\name{rbind_soundings}
\alias{rbind_soundings}
\title{Turn a list of soundings data frames into a single data frame}
\usage{
rbind_soundings(x)
}
\arguments{
\item{x}{a \code{list} of soundings \code{data.frame}s}
}
\value{
a "\code{rbind}ed" single data frame
}
\description{
A likely idiom used with this pacakge is to generate a sequence
of dates and then retrieve soundings data for those dates and
then turn them into one, big-ish data frame. Generating the
list of data frames is up to the caller but said list can then
be passed to this function to safely \code{\link[=rbind]{rbind()}} them into a data frame.
Empty results will be ignored.
}
