% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsound-package.R
\docType{package}
\name{unsound}
\alias{unsound}
\alias{unsound-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/unsound}
\item BugReports: \url{https://gitlab.com/hrbrmstr/unsound/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
